initial release
test for git ( Contains only main activity )
Note: android studio places a .git file within all android apps, but the only way for me to push git is to remove it. If we get errors in the development process regarding missing files or other odd build issues, remember that this could be it.